rename /Modal/Modal.tsx -> Modal/index.tsx
